backend/requirements.txt
# 网络小说智能问答系统 - Backend Dependencies
# Python 3.10+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# CORS & Middleware
python-jose[cryptography]==3.3.0

# AI & LLM
zhipuai==2.0.1
openai==1.54.0  # OpenAI & DeepSeek (兼容格式)
google-genai==1.52.0  # Google Gemini
dashscope>=1.19.0  # 阿里通义千问
langchain==0.1.0
langchain-community==0.0.10

# Vector Database
chromadb==0.4.22

# NLP
hanlp==2.1.0b55  # 中文NLP工具包，用于实体识别和分词
jieba==0.42.1    # 中文分词（BM25使用）
rank_bm25==0.2.2 # BM25算法

# Graph
networkx==3.2.1
scipy==1.11.4
# fa2==0.3.5  # 可选：ForceAtlas2 布局算法（Windows 上可能安装失败，可跳过）

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# File Parsing
ebooklib==0.18
chardet==5.2.0

# Visualization
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
cachetools==5.3.2  # 缓存工具

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.7.1
